Make perfdata output honour -b option
Status output show RAM usage as percentage, while perfdata output shows total literal ram usage.  If `-N -m` options are given, cached memory is *included* in both the plugin status output and the perfdata.  If `-N -b` options are given, buffered memory is *excluded* from the status output, but remains *included* in the perfdata.

This patch alters the behaviour of the `-N -b` option combination so that, if supplied, buffered memory is also *excluded* from the perfdata.